Add heartbeat handling for stability improvement
Added the logic of initiating a heartBeatTimer for client's and server's proper connection handling. Previously, it was initialized globally and invoked on each message send/receive which lacked management and caused issues with connection stability.

Now, heartbeat flagging is driven by a dedicated asio::steady_timer initiated per each ClientProcess for the server and Agent for the client. The timer is set to fire off after 3 minutes if no further network activity detected and implies breaking the current connection session. Consequently, this approach will lead to a more effective connection management and overall system stability.